Add delegate endpoint
