+++ backend/hct_mis_api/api/endpoints/rdi/__init__.py
@@ -4,8 +4,9 @@ from hct_mis_api.api.endpoints.rdi.base import (  # noqa: F401
     PushLaxToRDIView,
     PushToRDIView,
 )
-from hct_mis_api.api.endpoints.rdi.program import ProgramViewSet  # noqa: F401
-from hct_mis_api.api.endpoints.rdi.upload import UploadRDIView  # noqa: F401
-from hct_mis_api.api.endpoints.rdi.upload_people import (  # noqa: F401
-    UploadPeopleRDIView,
+from hct_mis_api.api.endpoints.rdi.delegate_people import (  # noqa: F401
+    DelegatePeopleRDIView,
 )
+from hct_mis_api.api.endpoints.rdi.program import ProgramViewSet  # noqa: F401
+from hct_mis_api.api.endpoints.rdi.push_people import PushPeopleToRDIView  # noqa: F401
+from hct_mis_api.api.endpoints.rdi.upload import UploadRDIView  # noqa: F401
